refactor: changed package structure
