+++ commitlint.config.js
@@ -1,6 +1,6 @@
 module.exports = {
   extends: ["@commitlint/config-conventional"],
   rules: {
-    "scope-enum": [2, "always", ["util", "node", "way", "relation", "changeset", "osmChange", "deps", "configurations"]],
+    "scope-enum": [2, "always", ["node", "way", "relation", "changeset", "osmChange", "deps", "configurations"]],
   },
 };
