+++ src/index.ts
@@ -1,57 +1,2 @@
-enum Types {
-  NODE = 'node',
-  WAY = 'way',
-  RELATION = 'relation',
-}
-interface BaseElement {
-  type: Types.NODE | Types.WAY | Types.RELATION;
-  id: number;
-  timestamp?: string;
-  version?: number;
-  changeset?: number;
-  user?: string;
-  tags?: { [key: string]: string };
-}
-
-interface OsmNode extends BaseElement {
-  type: Types.NODE;
-  lat: number;
-  lon: number;
-}
-
-interface OsmWay extends BaseElement {
-  type: Types.WAY;
-  nodes: OsmNode[];
-}
-
-interface OsmApiWay extends BaseElement {
-  type: Types.WAY;
-  nodes: number[];
-}
-
-type Member = BaseElement & {
-  role?: string;
-};
-
-interface OsmRelation extends BaseElement {
-  type: Types.RELATION;
-  members?: Member[];
-}
-
-interface Changeset {
-  type: 'changeset';
-  version?: string;
-  generator?: string;
-  tags?: { [key: string]: string };
-}
-
-interface OsmChange {
-  type: 'osmchange';
-  version?: string;
-  generator?: string;
-  create?: BaseElement[];
-  modify?: BaseElement[];
-  delete?: BaseElement[];
-}
-
-export { BaseElement, OsmNode, OsmWay, OsmApiWay, OsmRelation, Changeset, OsmChange, Member, Types };
+export * from './interfaces';
+export * from './util';
